docs: Add an FAQ
This is an initial attempt to get an FAQ up and running. I do not think that this is a permanent solution, but it should help us get the ball rolling on having an FAQ at all. I started this as suggested by @mistakia [here](https://github.com/orbitdb/orbit-db/issues/440#issuecomment-416347226).