Add a task to the notetaking process
Right now, the notetaker is on call for adding the notes to the repository. However, they should also post in the Gitter, for the simple reason that Zoom does not indicate if a call has happened or not. That means that any community members who show up late may wonder if the call is working, and may ask where the notes are if they are not yet in the repository. Posting a simple message - "Hey, call is over, here are the notes" - obviates this problem.