docs: Use orbitdb.maintainer field
The `maintainer` field turns out to be reserved, and not public, and it does not match the output of `npm info`. Npm does not offer a native way in the package.json to show who the maintainers are, bizarrely. In order to sidestep this, I propose here adding an orbitdb.maintainer field to our package.jsons. Using an "orbitdb" key allows us to future-proof this in case we want to use other metadata in the future in our npm manifests.